+++ packages/ui/src/components/Select.tsx
@@ -104,7 +104,7 @@ export const SelectSeparator = forwardRef<
 >(({ className, ...props }, ref) => (
   <SelectPrimitive.Separator
     ref={ref}
-    className={cn('-mx-1 my-1 h-px bg-surface-border', className)}
+    className={cn('-mx-1 my-1 h-px bg-line-neutral', className)}
     {...props}
   />
 ));